requirements.txt
python-dotenv
supabase
google-api-python-client
google-auth
google-auth-oauthlib
instagrapi
